refactor: 5 (facade removal) — retarget tests onto real modules
Repoint all `from cli.govkit import X` / `mod.X` / `govkit.X` test references
onto the modules that actually own those symbols (cli.marker, cli.fs,
cli.manifest, cli.cmd_*). test_govkit.py's `mod` handle is replaced with direct
symbol imports — a module handle collided with local `marker` dict variables.
With nothing importing through it anymore, the `X as X` re-export block in
govkit.py is deleted. govkit.py: 85 -> 68 lines and now imports only what it
uses (argparse + the 8 register fns). No behavior change; 807 tests green.
